+++ pkgs/cupertino_http/lib/src/cupertino_client.dart
@@ -15,6 +15,8 @@ import 'package:http/http.dart';
 
 import 'cupertino_api.dart';
 
+final _digitRegex = RegExp(r'^\d+$');
+
 class _TaskTracker {
   final responseCompleter = Completer<URLResponse>();
   final BaseRequest request;
@@ -279,6 +281,17 @@ class CupertinoClient extends BaseClient {
       throw ClientException('Redirect limit exceeded', request.url);
     }
 
+    final responseHeaders = response.allHeaderFields
+        .map((key, value) => MapEntry(key.toLowerCase(), value));
+
+    if (responseHeaders['content-length'] case final contentLengthHeader?
+        when !_digitRegex.hasMatch(contentLengthHeader)) {
+      throw ClientException(
+        'Invalid content-length header [$contentLengthHeader].',
+        request.url,
+      );
+    }
+
     return StreamedResponse(
       taskTracker.responseController.stream,
       response.statusCode,
@@ -288,8 +301,7 @@ class CupertinoClient extends BaseClient {
       reasonPhrase: _findReasonPhrase(response.statusCode),
       request: request,
       isRedirect: !request.followRedirects && taskTracker.numRedirects > 0,
-      headers: response.allHeaderFields
-          .map((key, value) => MapEntry(key.toLowerCase(), value)),
+      headers: responseHeaders,
     );
   }
 }